fix:add failpoints; divide code into files; batchID start from x-2; declare unsupported PK type; fix failpolicy bugs
